refactor: useCardType 커스텀 훅 삭제
