+++ hooks/src/lib/hooks/useCardNumber/useCardNumber.ts
@@ -1,20 +1,17 @@
 import { useState } from 'react';
-import useCardType from '../useCardType/useCardType';
-import { formatCardNumber } from '../../utils/card';
+import { formatCardNumber, getCardType } from '../../utils/card/card';
 import { getNumberErrorMessage, isNotNumber } from '../../utils/validation/validation';
+import type { CardType } from '../../utils/card/card.type';
 import { CARD_TYPE } from '../../constants/Condition';
-
-import type { CardType } from '../../types/common.type';
 
 const useCardNumber = (initialValue: string = '') => {
   const [cardNumber, setCardNumber] = useState(initialValue);
   const [isValidCardNumber, setIsValidCardNumber] = useState(false);
   const [cardNumberErrorMessage, setCardNumberErrorMessage] = useState('');
 
-  const { cardType, handleCardType }: { cardType: CardType; handleCardType: (cardNumber: string) => void } =
-    useCardType();
+  const [cardType, setCardType] = useState(getCardType(initialValue));
 
-  const validateCardNumber = (number: string) => {
+  const validateCardNumber = (cardType: CardType, number: string) => {
     const errorMessage = getNumberErrorMessage(number, CARD_TYPE[cardType].VALID_LENGTH);
 
     setCardNumberErrorMessage(errorMessage);
@@ -23,16 +20,16 @@ const useCardNumber = (initialValue: string = '') => {
 
   const handleCardNumberChange = (number: string) => {
     const cleanNumber = number.replace(/\s/g, '');
+    const newCardType = getCardType(cleanNumber);
 
-    handleCardType(cleanNumber);
+    if (cleanNumber.length > CARD_TYPE[newCardType].VALID_LENGTH) return;
 
-    if (cleanNumber.length > CARD_TYPE[cardType].VALID_LENGTH) return;
-
-    validateCardNumber(cleanNumber);
+    validateCardNumber(newCardType, cleanNumber);
 
     if (isNotNumber(cleanNumber)) return;
 
-    setCardNumber(formatCardNumber(cardType, cleanNumber));
+    setCardType(newCardType);
+    setCardNumber(formatCardNumber(newCardType, cleanNumber));
   };
 
   return {
